+++ pooltool/evolution/__init__.py
@@ -1,1 +1,3 @@
-from pooltool.evolution.event_based import simulate
+from pooltool.evolution.event_based.simulate import simulate as simulate_event_based
+
+simulate = simulate_event_based
